refactor: move Bedrock helpers to separate bedrock_utils.ts
Extract model mapping, streaming format conversion, and base64 decode
into dedicated module for better organization and testability.
